% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudokuAlt.R
\name{as.sudoku.matrix}
\alias{as.sudoku.matrix}
\title{as.sudoku.matrix}
\usage{
\method{as.sudoku}{matrix}(x, ...)
}
\arguments{
\item{x}{An n^2 x n^2 matrix}

\item{...}{other arguments (currently ignored)}
}
\value{
An object of class 'sudoku'
}
\description{
Construct a Sudoku Game Object
}
\details{
Coerce a matrix to an object that can be used as a sudoku game
}
\author{
Bill Venables
}

