% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudokuAlt.R
\name{as.sudoku.sudoku}
\alias{as.sudoku.sudoku}
\title{Construct a Sudoku Game Object}
\usage{
\method{as.sudoku}{sudoku}(x, ...)
}
\arguments{
\item{x}{A sudoku object}

\item{...}{other arguments (ignored)}
}
\value{
the input sudoku object
}
\description{
Identity function for sudoku objects
}
\author{
Bill Venables
}
