% Generated by roxygen2 (4.0.2): do not edit by hand
\name{daysAgo}
\alias{daysAgo}
\title{Format a Past Date}
\usage{
daysAgo(n = 0, warn = TRUE)
}
\arguments{
\item{n}{A positive integer for how many days ago}

\item{warn}{Issue a warning if n <= 0 or n > 30}
}
\value{
A character string of the form "dd/mm/yy"
}
\description{
Format a Date Relative to the Current Date
}
\details{
Internal function used by fetchUKGame().
}
\examples{
daysAgo()  ## today
daysAgo(7) ## a week ago
}
\author{
Bill Venables
}

