% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudokuAlt.R
\name{emptyGame}
\alias{emptyGame}
\title{Construct an empty game}
\usage{
emptyGame(n = 3)
}
\arguments{
\item{n}{an integer value between 2 and 5 inclusive.}
}
\value{
An empty sudoku game structure
}
\description{
Construct a Vacant Game Structure
}
\details{
Returns a vacant game structure to allow special patterns to
be constructed, as shown in the example.
}
\examples{
g <- emptyGame(4)
diag(g) <- LETTERS[1:16]
g \%>\% solve \%>\% plot -> sg ## \%>\% imported from magrittr
}
\author{
Bill Venables
}
