% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetchUKGame}
\alias{fetchUKGame}
\title{Retrieve a Sudoku from the UK Site}
\usage{
fetchUKGame(day = NULL)
}
\arguments{
\item{day}{positive integer < 30, how many days ago? or NULL for
the current day}
}
\value{
The published sudoku game
}
\description{
Retrieve a Sudoku Game
}
\details{
Connects to http://www.sudoku.org.uk/DailySudoku.asp and retrieves
the sudoku game from 'day' days ago.  Based on a function from a
related sudoku package, sudoku::fetchSudokuUK with minor changes.
}
\examples{
\dontrun{
(g0 <- fetchUKGame())  ## today's game
(g3 <- fetchUKGame(3)) ## game from 3 days ago
}
}
\author{
Bill Venables
}

