% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeGame}
\alias{makeGame}
\title{Make a New Sudoku Game}
\usage{
makeGame(n = 3, gaps = ceiling(2 * n^4/3), maxit = 5)
}
\arguments{
\item{n}{Size of the game, n^2 x n^2}

\item{gaps}{Number of holes to leave for the solution}

\item{maxit}{Number of tries before giving up.}
}
\value{
a sudoku game
}
\description{
Construct a Random Sudoku Game
}
\details{
Construcs a sudoku game for given n, 2 <= n <= 5.
n = 5 can be problematical.
}
\examples{
set.seed(54321)
(m <- makeGame())
sm <- solve(m)
plot(sm)
}
\author{
Bill Venables
}

