% Generated by roxygen2 (4.0.2): do not edit by hand
\name{originalGame}
\alias{originalGame}
\title{Retrieve the Original from a Solved Game}
\usage{
originalGame(x)
}
\arguments{
\item{x}{a sudoku object}
}
\value{
The original sudoku game corresponding to the solution,
or object itself if the game is unsolved
}
\description{
Retrieve the Original from a Solved Game
}
\details{
Convenience function for accessing an original from a solved game.
If the game is unsolved, the object itself is returned.
}
\examples{
set.seed(666)
(sg <- solve(seedGame()))
originalGame(sg)
}
\author{
Bill Venables
}

