% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.sudoku}
\alias{plot.sudoku}
\title{Plot a Sudoku Game}
\usage{
\method{plot}{sudoku}(x, ..., cex = 2 - (n - 3)/2, colSolution = "grey",
  colGame = "fire brick")
}
\arguments{
\item{x}{The sudoku game}

\item{...}{additional arguments}

\item{cex}{Character expansion factor}

\item{colSolution}{colour to be used for the solution (if present)}

\item{colGame}{colour to be used for the original game}
}
\description{
Plot a Sudoku Game
}
\details{
Present a graphical display of a sudoku game and its solution if
the game is solved
}
\examples{
set.seed(1234)
plot(solveGame(seedGame(4)))
}
\author{
Bill Venables
}

