% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sudokuAlt.R
\name{seedGame}
\alias{seedGame}
\title{Starting Point to Make a Random Sudoku Game}
\usage{
seedGame(n = 3)
}
\arguments{
\item{n}{Size of the game, n^2 x n^2}
}
\value{
A sparse unsolved sudoku game
}
\description{
Generate a random sudoku game starting point
}
\details{
Generates a game with one instance of each symbol in random
positions.
}
\examples{
set.seed(2345)
g <- seedGame(3)
sg <- solve(g) ## a completed random game
plot(sg)
}
\author{
Bill Venables
}

