% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudokuAlt.R
\name{makeGame}
\alias{makeGame}
\title{Make a New Sudoku Game}
\usage{
makeGame(n = 3, gaps = ceiling(3 * n^4/4), maxit = 25)
}
\arguments{
\item{n}{Size of the game, n^2 x n^2}

\item{gaps}{Number of holes to leave for the solution}

\item{maxit}{Number of tries before giving up.}
}
\value{
a sudoku game
}
\description{
Construct a Random Sudoku Game
}
\details{
Construcs a sudoku game for given n, 2 <= n <= 5.
n = 5 can be problematical.
}
\examples{
set.seed(54321)
makeGame() \%>\% solve \%>\% plot -> sg
originalGame(sg)

}
\author{
Bill Venables
}
