% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudokuAlt.R
\name{plot.sudoku}
\alias{plot.sudoku}
\title{Plot a Sudoku Game}
\usage{
\method{plot}{sudoku}(
  x,
  ...,
  cex = 1.5 - (n - 3)/2,
  colSolution = "grey",
  colGame = "fire brick"
)
}
\arguments{
\item{x}{The sudoku game}

\item{...}{additional arguments}

\item{cex}{Character expansion factor}

\item{colSolution}{colour to be used for the solution (if present)}

\item{colGame}{colour to be used for the original game}
}
\value{
The sudoku game \code{x}, invisibly.
}
\description{
Plot a Sudoku Game
}
\details{
Present a graphical display of a sudoku game and its solution if
the game is solved
}
\examples{
set.seed(20191)
makeGame(4, gaps = 0) \%>\% plot(cex=1) -> sg
}
\author{
Bill Venables
}
