% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudokuAlt.R
\name{print.sudoku}
\alias{print.sudoku}
\title{Print a Sudoku Object}
\usage{
\method{print}{sudoku}(x, ...)
}
\arguments{
\item{x}{The sudoku game object}

\item{...}{extra arguments (ignored)}
}
\value{
the object, invisibly
}
\description{
Print a Sudoku Object
}
\details{
Prints a sudoku object in an easily recognisable form.
}
\author{
Bill Venables
}
